Remove execution date reference in docs
Not all usages can be removed, unfortunately, since there are still many
usages requiring the value, especially in Python APIs.

This also adds additional values for Sentry tags and Elastic log
filenames based on data interval instead of execution date, so the user
can migrate as soon as possible, even if we have not changed the default.